requirements.txt
Flask
Flask-Restless
Flask-Restful
Flask-SQLAlchemy
Flask-Script
pymysql
pyjwt